requirements.txt
MetaTrader5
pandas
numpy
torch
ta
joblib
python-dotenv
scikit-learn
tqdm
